По замечаниям. Уберет асинхронную цепочку из index. Упростит App. При отсутствии страницы, по умолчанию загружается Main. Подколючит Card к redux. LocationsList получает проп city, а не использует внутренний state. Обновит тесты.
