+++ src/components/card/card.e2e.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, {shallow} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import Card from "./card.jsx";
+import {Card} from "./card.jsx";
 
 import offers from "../../mocks/tests_offers.js";
