+++ src/components/card/card.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import Card from "./card.jsx";
+import {Card} from "./card.jsx";
 
 import offers from "../../mocks/tests_offers.js";
