+++ src/components/cards-list/cards-list.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import Card from "../card/card.jsx";
 
 const CardsList = (props) => {
-  const {offers, isNearPlaces = false, onPlaceCardHeaderClick, onActiveItemChange, onFavoritesToggle} = props;
+  const {offers, isNearPlaces, onPlaceCardHeaderClick, onActiveItemChange} = props;
 
   return (
     offers.map((offer) => {
@@ -14,7 +14,6 @@ const CardsList = (props) => {
           isNearPlaces={isNearPlaces}
           onPlaceCardHeaderClick={onPlaceCardHeaderClick}
           onActiveItemChange={onActiveItemChange}
-          onFavoritesToggle={onFavoritesToggle}
         />
       );
     })
@@ -36,7 +35,6 @@ CardsList.propTypes = {
   ),
   isNearPlaces: PropTypes.bool,
   onActiveItemChange: PropTypes.func.isRequired,
-  onFavoritesToggle: PropTypes.func.isRequired,
 };
 
 export default CardsList;
