+++ src/components/locations-list/locations-list.jsx
@@ -3,8 +3,7 @@ import PropTypes from "prop-types";
 
 class LocationsList extends React.PureComponent {
   render() {
-    const {locations, onCityChange, onActiveItemChange, activeItemId} = this.props;
-    const city = activeItemId;
+    const {locations, onCityChange, city} = this.props;
 
     return (
       <ul className="locations__list tabs__list">
@@ -14,7 +13,6 @@ class LocationsList extends React.PureComponent {
               <a
                 onClick={(evt) => {
                   evt.preventDefault();
-                  onActiveItemChange(it);
                   onCityChange(it);
                 }}
                 className={`locations__item-link tabs__item${it === city ? ` tabs__item--active` : ``}`}
@@ -32,9 +30,8 @@ class LocationsList extends React.PureComponent {
 
 LocationsList.propTypes = {
   locations: PropTypes.arrayOf(PropTypes.string).isRequired,
-  activeItemId: PropTypes.any.isRequired,
+  city: PropTypes.string.isRequired,
   onCityChange: PropTypes.func.isRequired,
-  onActiveItemChange: PropTypes.func.isRequired,
 };
 
 export default LocationsList;
