+++ src/components/main/main.jsx
@@ -2,22 +2,17 @@ import React from "react";
 import PropTypes from "prop-types";
 import {connect} from "react-redux";
 
-import {ActionCreator, Operation} from "../../reducer/data/data.js";
+import {ActionCreator} from "../../reducer/data/data.js";
 import {getFilteredOffers, getLocations, getCity} from "../../reducer/data/selectors.js";
 import {getSortType} from "../../reducer/app/selectors.js";
-import history from "../../history.js";
-import {AppRoute} from "../../const.js";
 
-import withActiveItem from "../../hocs/with-active-item/with-active-item.js";
 import Header from "../header/header.jsx";
 import LocationsList from "../locations-list/locations-list.jsx";
 import Places from "../places/places.jsx";
 import PlacesEmpty from "../places-empty/places-empty.jsx";
 
-const LocationsListWrapped = withActiveItem(LocationsList);
-
 const Main = (props) => {
-  const {onPlaceCardHeaderClick, city, activeOffers, locations, onCityChange, onActiveItemChange, activeItemId, sortType, onFavoritesToggle} = props;
+  const {onChangeActiveOfferId, city, activeOffers, locations, onCityChange, onActiveItemChange, activeItemId, sortType} = props;
 
   return (
     <div className={`page page--gray page--main${activeOffers.length ? `` : ` page__main--index-empty`}`}>
@@ -28,8 +23,8 @@ const Main = (props) => {
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
           <section className="locations container">
-            <LocationsListWrapped
-              initActiveItemId={locations[0]}
+            <LocationsList
+              city={city}
               locations={locations}
               onCityChange={onCityChange}
             />
@@ -40,12 +35,11 @@ const Main = (props) => {
             ?
             <Places
               activeOffers={activeOffers}
-              onPlaceCardHeaderClick={onPlaceCardHeaderClick}
+              onPlaceCardHeaderClick={onChangeActiveOfferId}
               onActiveItemChange={onActiveItemChange}
               activeItemId={activeItemId}
               city={city}
               sortType={sortType}
-              onFavoritesToggle={onFavoritesToggle}
             />
             :
             <PlacesEmpty />
@@ -57,7 +51,7 @@ const Main = (props) => {
 };
 
 Main.propTypes = {
-  onPlaceCardHeaderClick: PropTypes.func.isRequired,
+  onChangeActiveOfferId: PropTypes.func.isRequired,
   city: PropTypes.string.isRequired,
   activeOffers: PropTypes.array.isRequired,
   activeItemId: PropTypes.any.isRequired,
@@ -65,7 +59,6 @@ Main.propTypes = {
   onCityChange: PropTypes.func.isRequired,
   onActiveItemChange: PropTypes.func.isRequired,
   sortType: PropTypes.string.isRequired,
-  onFavoritesToggle: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = (state) => {
@@ -81,14 +74,9 @@ const mapDispatchToProps = (dispatch) => ({
   onCityChange(city) {
     dispatch(ActionCreator.changeCity(city));
   },
-  onFavoritesToggle(offerId, favoriteStatus) {
-    dispatch(Operation.postFavorite(offerId, favoriteStatus))
-    .catch((error) => {
-      if (error.response.status === 401) {
-        history.push(AppRoute.LOGIN);
-      }
-    });
-  },
+  onChangeActiveOfferId(id) {
+    dispatch(ActionCreator.changeActiveOfferId(id));
+  }
 });
 
 export {Main};
